Move sanitization in the same place
