+++ src/Form/Fields/SharpFormTextareaField.php
@@ -3,16 +3,17 @@
 namespace Code16\Sharp\Form\Fields;
 
 use Code16\Sharp\Form\Fields\Formatters\TextareaFormatter;
-use Code16\Sharp\Form\Fields\Utils\SharpFormFieldWithHtmlSanitization;
 use Code16\Sharp\Form\Fields\Utils\SharpFormFieldWithMaxLength;
 use Code16\Sharp\Form\Fields\Utils\SharpFormFieldWithPlaceholder;
 use Code16\Sharp\Utils\Fields\IsSharpFieldWithLocalization;
 use Code16\Sharp\Utils\Fields\SharpFieldWithLocalization;
+use Code16\Sharp\Utils\Sanitization\IsSharpFieldWithHtmlSanitization;
+use Code16\Sharp\Utils\Sanitization\SharpFieldWithHtmlSanitization;
 
-class SharpFormTextareaField extends SharpFormField implements IsSharpFieldWithLocalization
+class SharpFormTextareaField extends SharpFormField implements IsSharpFieldWithHtmlSanitization, IsSharpFieldWithLocalization
 {
+    use SharpFieldWithHtmlSanitization;
     use SharpFieldWithLocalization;
-    use SharpFormFieldWithHtmlSanitization;
     use SharpFormFieldWithMaxLength;
     use SharpFormFieldWithPlaceholder;
