+++ src/Utils/Sanitization/IsSharpFieldWithHtmlSanitization.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Code16\Sharp\Utils\Sanitization;
+
+interface IsSharpFieldWithHtmlSanitization
+{
+    public function isSanitizingHtml(): bool;
+}
